% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{col_normalize}
\alias{col_normalize}
\title{Normalize Columns of a Matrix}
\usage{
col_normalize(matrix)
}
\arguments{
\item{matrix}{A numeric matrix to be normalized}
}
\value{
A matrix with columns normalized by their sums
}
\description{
Normalize Columns of a Matrix
}
\keyword{internal}
