% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{common_signature_filter}
\alias{common_signature_filter}
\title{Filter Signatures for Common Nodes Across Graphs}
\usage{
common_signature_filter(graphs, signatures)
}
\arguments{
\item{graphs}{A list of igraph objects}

\item{signatures}{A list of node signatures (vectors of node names)}
}
\value{
A list of filtered signatures containing only nodes common to all graphs
}
\description{
This function takes a list of graphs and a list of node signatures, and filters
each signature to include only nodes that are present in the largest connected
component of all provided graphs.
}
\details{
For each signature:
\enumerate{
\item It checks which nodes are in the largest connected component of each graph.
\item It finds the intersection of these nodes across all graphs.
\item It returns this intersection as the new filtered signature.
}
}
\keyword{internal}
