% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{consensus_net}
\alias{consensus_net}
\title{Create a Consensus Network from Multiple Networks}
\usage{
consensus_net(list_of_nets, threshold = 0.9)
}
\arguments{
\item{list_of_nets}{A list of adjacency matrices, each representing a network.}

\item{threshold}{A numeric value between 0 and 1 representing the consensus threshold. Default is 0.9.}
}
\value{
An adjacency matrix representing the consensus network.
}
\description{
This function creates a consensus network from a list of adjacency matrices by averaging
and thresholding the connections.
}
\details{
The function performs the following steps:
\enumerate{
\item Averages all input adjacency matrices.
\item Applies a threshold to the averaged matrix.
\item Returns a binary adjacency matrix where 1 indicates a consensus connection and 0 indicates no consensus.
}
}
\keyword{internal}
