% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_methods.R
\name{correlated_sigs}
\alias{correlated_sigs}
\title{Recontextualize seed signatures with correlation based neighbors}
\usage{
correlated_sigs(corr_mat, seeds, limit = 30)
}
\arguments{
\item{corr_mat}{Correlation Matrix}

\item{seeds}{Either a single unnamed gene "TP53", a named list of genes, or a list of named lists of genes.}

\item{limit}{Number of genes to keep in the output, or a vector of lengths. Default is 30.}
}
\description{
Recontextualize seed signatures with correlation based neighbors
}
\keyword{internal}
