% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drugmatrix.liver}
\alias{drugmatrix.liver}
\title{Drug Matrix Liver Dataset}
\format{
Nested list of genesets
}
\source{
\url{https://ntp.niehs.nih.gov/data/drugmatrix}
}
\usage{
drugmatrix.liver
}
\description{
DEGs between drug (max dose) and control samples in DrugMatrix liver data.
There are 57 drugs and  DEGs are mouse genes.
For each geneset, the \code{up} set represents the top 50 DEGs and the \code{up_full} represents the top 1000 DEGs.
}
\keyword{datasets}
