% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{eset_stratified_sample}
\alias{eset_stratified_sample}
\title{Perform Stratified Sampling on an ExpressionSet}
\usage{
eset_stratified_sample(eset, stratify_column, sample_id_column, p = 0.5)
}
\arguments{
\item{eset}{An ExpressionSet object containing gene expression data and phenotype data.}

\item{stratify_column}{The name of the column in pData(eset) to use for stratification.}

\item{sample_id_column}{The name of the column in pData(eset) containing sample IDs.}

\item{p}{The proportion of samples to include in the training set. Default is 0.5.}
}
\value{
A list containing two ExpressionSet objects:
\item{train}{The ExpressionSet containing the training samples}
\item{test}{The ExpressionSet containing the test samples}
}
\description{
This function performs stratified sampling on an ExpressionSet object, splitting it into training and test sets.
}
\details{
The function stratifies the samples based on the specified column and then randomly selects
a proportion p of samples from each stratum for the training set. The remaining samples
form the test set.
}
