% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_signature}
\alias{get_signature}
\title{Extract Signature Nodes from Random Walk Results}
\usage{
get_signature(rwr_df, sig_label, node_label, cancer_label)
}
\arguments{
\item{rwr_df}{A dataframe containing the results of RWR, including columns for
cancer type, signature, node type, and node names.}

\item{sig_label}{The label of the signature to filter for.}

\item{node_label}{The label of the node type to filter for.}

\item{cancer_label}{The label of the cancer type to filter for.}
}
\value{
A vector of node names that are significant for the specified signature,
node type, and cancer type.
}
\description{
Extract Signature Nodes from Random Walk Results
}
