% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex.blood.aging}
\alias{gtex.blood.aging}
\title{GTEX Whole Blood Aging signatures (Continuous)}
\format{
Nested list of genesets
}
\source{
\url{https://gtexportal.org/home/downloads/adult-gtex/bulk_tissue_expression}
}
\usage{
gtex.blood.aging
}
\description{
DEGs between old and young whole blood bulk samples (age as a numerical covariate) from GTEX v10.
For each geneset, the \code{up} set represents the top 100 DEGs
and the \code{up_full} represents the full ranked list of 23131 genes ranked by logFC*-log10(adj.p.val).
}
\keyword{datasets}
