% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex.brain.aging.20.70}
\alias{gtex.brain.aging.20.70}
\title{GTEX Whole Brain Aging signatures (Discrete Bins)}
\format{
Nested list of genesets
}
\source{
\url{hhttps://gtexportal.org/home/downloads/adult-gtex/bulk_tissue_expression}
}
\usage{
gtex.brain.aging.20.70
}
\description{
DEGs between 70-79 and 20-29 brain bulk samples from GTEX v10.
For each geneset, the \code{up} set represents the top 100 DEGs
and the \code{up_full} represents the full ranked list of 23131 genes ranked by logFC*-log10(adj.p.val).
}
\keyword{datasets}
