% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{jaccard}
\alias{jaccard}
\title{Calculate Jaccard Similarity Between Two Sets}
\usage{
jaccard(a, b)
}
\arguments{
\item{a}{A vector representing a set}

\item{b}{A vector representing a set}
}
\value{
The Jaccard similarity score between sets a and b
}
\description{
Calculate Jaccard Similarity Between Two Sets
}
