% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_methods.R
\name{network_sig}
\alias{network_sig}
\title{Network-propagation based Recontextualization.}
\usage{
network_sig(
  ig,
  seeds,
  sig = c("corr", "rwr"),
  avg_p = FALSE,
  avg_p_vals = c(1e-04, 0.1),
  avg_p_length = 5,
  p = 0.1,
  bootstrap = FALSE,
  n_bootstraps = 1000,
  limit = 30
)
}
\arguments{
\item{ig}{network given as an igraph}

\item{seeds}{Either a single unnamed gene "TP53", a named list of genes, or a list of named lists of genes.}

\item{sig}{A string specifying the type of network signature: random walk, correlation etc.}

\item{avg_p}{A boolean specifying whether to ensemble random walk results over a range of restart values}

\item{avg_p_vals}{A numeric vector specifying the start and end of a arithmetic sequence to explore restart values.}

\item{avg_p_length}{A numeric specifying how many values within \code{avg_p_vals} to include in the ensemble}

\item{p}{A numeric specifying the restart value for random walk, default=0.1}

\item{bootstrap}{A boolean specifying whether to use empirical distributions of stationary values to find significant genes.}

\item{n_bootstraps}{A numeric specifying the number of bootstraps to perform.}

\item{limit}{A numeric specifying the number of genes to be included in the network signature. Default is 30.}
}
\value{
vector of gene strings
}
\description{
Network-propagation based Recontextualization.
}
