% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neurips2023.t}
\alias{neurips2023.t}
\title{Neurips 2023 T Cell signatures}
\format{
Nested list of genesets
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE279945}
}
\usage{
neurips2023.t
}
\description{
DEGs between drug and control samples (DMSO) in human T cells.
There are 135 shared drugs across the 4 cell types and DEGs are HGNC symbols.
For each geneset, the \code{up} set represents the top 100 DEGs
and the \code{up_full} represents the full ranked list of 6349 genes ranked by logFC*-log10(adj.p.val).
The pseudobulk dataset these genesets were derived from had 6349 genes.
}
\keyword{datasets}
