% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{perturbseq.rpe1}
\alias{perturbseq.rpe1}
\title{Perturb-seq RPE1 Dataset}
\format{
Nested list of genesets
}
\source{
\url{https://plus.figshare.com/articles/dataset/_Mapping_information-rich_genotype-phenotype_landscapes_with_genome-scale_Perturb-seq_Replogle_et_al_2022_processed_Perturb-seq_datasets/20029387}
}
\usage{
perturbseq.rpe1
}
\description{
DEGs between drug (max dose) and control samples in SciPlex RPE1 cell line data.
There are 585 CRISPRi knockdowns and DEGs are HGNC symbols.
For each geneset, the \code{up} set represents the top 50 DEGs and the \code{up_full} represents the top 1000 DEGs.
}
\keyword{datasets}
