% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_methods.R
\name{random_walk}
\alias{random_walk}
\title{Perform a random walk with restart (personalized page rank) on an igraph given a seed matrix, and return stationary probabilties.
Stripped down and corrected version of dnet: https://rdrr.io/cran/dnet/src/R/dRWR.r}
\usage{
random_walk(
  ig,
  seed_mat,
  restart = 0.1,
  epsilon = NULL,
  normalize = c("row", "column", "laplacian", "none")
)
}
\arguments{
\item{ig}{igraph object}

\item{seed_mat}{(Gene, num_seeds) matrix with prior weights for each gene in a seed set. See seed_matrix.}

\item{restart}{the restart probability for RWR}

\item{epsilon}{Exploration factor}

\item{normalize}{Normalization strategy}
}
\value{
It returns a sparse matrix with stationary probabilities.
}
\description{
Perform a random walk with restart (personalized page rank) on an igraph given a seed matrix, and return stationary probabilties.
Stripped down and corrected version of dnet: https://rdrr.io/cran/dnet/src/R/dRWR.r
}
\keyword{internal}
