% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_evaluation.R
\name{recon_eval_df}
\alias{recon_eval_df}
\title{Network Propagation Helper}
\usage{
recon_eval_df(
  ig,
  seed_name,
  source_sigs,
  dest_sigs,
  restart = 0.75,
  avg_p = FALSE,
  avg_p_vals = c(1e-04, 0.1),
  avg_p_length = 5,
  bootstrap = FALSE,
  n_bootstraps = 1000,
  recon = TRUE,
  use_weights = TRUE,
  weights.pwr = 1,
  normalize = c("row", "column", "laplacian"),
  save = FALSE,
  save_path = "",
  limit = 30
)
}
\arguments{
\item{ig}{igraph}

\item{seed_name}{name of list of genesets to be used to label dataframe}

\item{source_sigs}{list of genesets (to be recontextualized)}

\item{dest_sigs}{list of genesets (the ground truth)}

\item{restart}{restart value}

\item{avg_p}{A boolean specifying whether to ensemble random walk results over a range of restart values}

\item{avg_p_vals}{A numeric vector specifying the start and end of a geometric sequence to explore restart values.}

\item{avg_p_length}{A numeric specifying how many values within \code{avg_p_vals} to include in the ensemble}

\item{bootstrap}{A boolean specifying whether to use empirical distributions of stationary values to find significant genes.}

\item{n_bootstraps}{A numeric specifying the number of bootstraps to perform.}

\item{recon}{Boolean indicating whether recontextualization occurs}

\item{use_weights}{Boolean indicating whether to use weights in the KS Test}

\item{weights.pwr}{Power to raise weights to}

\item{normalize}{Normalization strategy to employ}

\item{save}{Boolean indicating whether to save recontextualized signatures}

\item{save_path}{file path to save recontextualized signatures}

\item{limit}{Number of genes to keep in the output, or a vector of lengths}
}
\description{
Network Propagation Helper
}
\keyword{internal}
