% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_methods.R
\name{rwr_mat}
\alias{rwr_mat}
\title{Perform a random walk on an igraph given seeds, return stationary probabilities}
\usage{
rwr_mat(
  ig,
  seeds,
  restart = 0.75,
  epsilon = NULL,
  normalize = c("row", "column", "laplacian", "none")
)
}
\arguments{
\item{ig}{igraph object}

\item{seeds}{Gene Symbol(s) that are seed nodes for the random walk. If random walk is to be done on multiple genes individually or multiple sets of genes, they should be in seperate elements of the list.}

\item{restart}{Numeric, Probability of restarting at the seed nodes}

\item{epsilon}{Exploration factor}

\item{normalize}{Normalization strategy}
}
\value{
(n_gene, n_seeds) matrix of stationary probability values
}
\description{
Perform a random walk on an igraph given seeds, return stationary probabilities
}
