% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sciplex.k562}
\alias{sciplex.k562}
\title{Sciplex K562 Dataset}
\format{
Nested list of genesets
}
\source{
\url{https://figshare.com/articles/dataset/sciPlex_dataset/24681285?file=43381398}
}
\usage{
sciplex.k562
}
\description{
DEGs between drug (max dose) and control samples in SciPlex K562 cell line data.
There are 60 drugs and DEGs are EnsemblIDs.
For each geneset, the \code{up} set represents the top 100 DEGs
and the \code{up_full} represents the full ranked list of genes ranked by logFC*-log10(adj.p.val).
The length of \code{up_full} ranges from 2778 - 3776 based on the perturbation.
}
\keyword{datasets}
