% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_methods.R
\name{seed_matrix}
\alias{seed_matrix}
\title{Create (gene x seed) prior matrix based on seed signatures.}
\usage{
seed_matrix(ig, seeds, bootstrap = FALSE, n_bootstraps = 1000)
}
\arguments{
\item{ig}{An igraph object representing the network that has the seed genes as vertices}

\item{seeds}{Either a single unnamed gene "TP53", a named list of genes, or a list of named lists of genes.}

\item{bootstrap}{A boolean specifying whether to use empirical distributions of stationary values to find significant genes.}

\item{n_bootstraps}{A numeric specifying the number of bootstraps to perform.}
}
\value{
A binary matrix where rows represent genes in the graph and columns represent seed sets.
}
\description{
This function creates a binary matrix representing seed genes in the context of a graph.
}
\keyword{internal}
