% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{split_learn_save}
\alias{split_learn_save}
\title{Split Dataset, Learn Network, and Save Results}
\usage{
split_learn_save(
  eset,
  stratify_col,
  sample_id_col,
  path,
  name,
  igraph = FALSE,
  diag_zero = FALSE,
  p = 0.5,
  n_split = 10,
  cor.type = c("unsigned", "signed hybrid", "signed"),
  learn = c("wgcna", "megena", "aracne", "shine")
)
}
\arguments{
\item{eset}{An ExpressionSet object containing gene expression data.}

\item{stratify_col}{Column name in eset's pData for stratification.}

\item{sample_id_col}{Column name in eset's pData for sample IDs.}

\item{path}{Directory path to save the results.}

\item{name}{Base name for the output files.}

\item{igraph}{If TRUE, returns igraph objects instead of matrices. Default is FALSE.}

\item{diag_zero}{If TRUE, sets the diagonal of adjacency matrices to zero. Default is FALSE.}

\item{p}{Proportion of samples to include in the training set. Default is 0.5.}

\item{n_split}{Number of times to repeat the split-learn-save process. Default is 10.}

\item{cor.type}{Type of correlation for WGCNA. Options are "unsigned", "signed hybrid", or "signed". Default is "unsigned".}

\item{learn}{Learning method to use. Currently only "wgcna" is implemented. Default is "wgcna".}
}
\description{
This function splits an expression set, learns a gene network using a specified method,
and saves the results for both training and test sets.
}
\details{
For each split:
\enumerate{
\item The eset is split into training and test sets using stratified sampling.
\item A network is learned on both the training and test sets using the specified method.
\item The resulting networks are saved as RDS files.
}
}
\note{
This function requires the eset_stratified_sample and wgcna.adj functions to be available.
Currently, only the WGCNA method is fully implemented.
}
\examples{
\dontrun{
library(Biobase)
data(sample.ExpressionSet)
split_learn_save(sample.ExpressionSet, "type", "sample", "output_dir", "network")
}

}
