% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tahoe.lovo}
\alias{tahoe.lovo}
\title{Tahoe LOVO Dataset}
\format{
Nested list of genesets
}
\source{
\url{https://github.com/ArcInstitute/arc-virtual-cell-atlas/tree/main/tahoe-100M}
}
\usage{
tahoe.lovo
}
\description{
DEGs between drug and control samples in Tahoe LOVO cell line data.
There are 109 drugs and gene identifiers are a mixture of HGNC symbols and Ensembl IDs.
For each geneset, the \code{up} set represents the top 100 DEGs
and the \code{up_full} represents the full ranked list of genes ranked by logFC*-log10(adj.p.val).
The length of \code{up_full} ranges from 3968 - 27303 based on the perturbation.
}
\keyword{datasets}
