% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_methods.R
\name{top_n_mat}
\alias{top_n_mat}
\title{Find the top n genes per seed in a (gene x seed) matrix.
This is the recontextualized signature.
If doing ks.test, you don't need to find the top_n. Just find ks.test(original, recontextualized ranking) before and after.}
\usage{
top_n_mat(mat, limit = 30)
}
\arguments{
\item{mat}{(n_gene, n_seed) matrix of stationary probability values from rwr_mat}

\item{limit}{Number of genes to keep in the output, or a vector of lengths. Default is 30.}
}
\value{
A named list of genesets. Each list element is the recontextualized signature for that seed.
}
\description{
Find the top n genes per seed in a (gene x seed) matrix.
This is the recontextualized signature.
If doing ks.test, you don't need to find the top_n. Just find ks.test(original, recontextualized ranking) before and after.
}
