% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_evaluation.R
\name{vcorrelated_sigs}
\alias{vcorrelated_sigs}
\title{Recontextualize seed signatures with correlation based neighbors}
\usage{
vcorrelated_sigs(corr_mats, seeds, limit = 30)
}
\arguments{
\item{corr_mats}{List of correlation matrices}

\item{seeds}{List of seeds, length has to match corr_mats}

\item{limit}{Numeric indicating the number of genes to be returned}
}
\description{
Recontextualize seed signatures with correlation based neighbors
}
