% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vjaccard}
\alias{vjaccard}
\title{Calculate Jaccard Similarity for Multiple Pairs of Sets}
\usage{
vjaccard(sigs1, sigs2)
}
\arguments{
\item{sigs1}{A list of sets (vectors)}

\item{sigs2}{A list of sets (vectors), must be the same length as sigs1}
}
\value{
A vector of Jaccard similarity scores
}
\description{
Calculate Jaccard Similarity for Multiple Pairs of Sets
}
