% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{wgcna.adj}
\alias{wgcna.adj}
\title{Construct WGCNA Adjacency Matrix}
\usage{
wgcna.adj(
  mat,
  min.sft = 0.85,
  beta = NULL,
  cores = 1,
  cor.fn = c("bicor", "cor"),
  cor.type = c("unsigned", "signed hybrid", "signed"),
  powers = c(seq(1, 10, by = 1), seq(12, 20, by = 2)),
  igraph = FALSE,
  diag_zero = FALSE
)
}
\arguments{
\item{mat}{A matrix with rows as samples and columns as genes}

\item{min.sft}{Minimum scale-free topology fitting index R^2 to pick soft-thresholding power. Default is 0.85.}

\item{beta}{Soft-thresholding power. If NULL, it will be automatically selected. Default is NULL.}

\item{cores}{Number of CPU cores to use for parallel computing. Default is 1.}

\item{cor.fn}{Correlation function to use. Either "bicor" (biweight midcorrelation) or "cor" (Pearson correlation). Default is "bicor".}

\item{cor.type}{Type of correlation network. Options are "unsigned", "signed hybrid", or "signed". Default is "unsigned".}

\item{powers}{Vector of soft-thresholding powers to try. Default is c(seq(1, 10, by = 1), seq(12, 20, by = 2)).}

\item{igraph}{If TRUE, returns an igraph object instead of a matrix. Default is FALSE.}

\item{diag_zero}{If TRUE, sets the diagonal of the adjacency matrix to zero. Default is FALSE.}
}
\value{
An adjacency matrix or an igraph object representing the gene co-expression network.
}
\description{
This function constructs a weighted gene co-expression network adjacency matrix using WGCNA.
}
\details{
This function performs the following steps:
\enumerate{
\item Prepares the expression data.
\item Selects the soft-thresholding power (if not provided).
\item Constructs the adjacency matrix using WGCNA::adjacency().
\item Optionally converts the result to an igraph object.
}
}
