% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{wgcna.power}
\alias{wgcna.power}
\title{Construct WGCNA Adjacency Matrix from Correlation Matrix}
\usage{
wgcna.power(cor_mat, cores = 1, diag_zero = TRUE)
}
\arguments{
\item{cor_mat}{A correlation matrix of gene expression data.}

\item{cores}{Number of CPU cores to use for parallel computing. Default is 1.}

\item{diag_zero}{If TRUE, sets the diagonal of the adjacency matrix to zero. Default is TRUE.}
}
\value{
An adjacency matrix representing the gene co-expression network.
}
\description{
This function constructs a weighted gene co-expression network adjacency matrix using WGCNA,
starting from a pre-computed correlation matrix.
}
\details{
This function performs the following steps:
\enumerate{
\item Selects the optimal soft-thresholding power using WGCNA::pickSoftThreshold.fromSimilarity().
\item Constructs the adjacency matrix using WGCNA::adjacency.fromSimilarity().
\item Optionally sets the diagonal to zero and/or converts the result to an igraph object.
}
}
\keyword{internal}
