% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_methods.R
\name{annotate_prob_vec}
\alias{annotate_prob_vec}
\title{Returns a dataframe object with the stationary probability value and column indicating whether gene was a seed}
\usage{
annotate_prob_vec(prob_vec, seeds)
}
\arguments{
\item{prob_vec}{(n_gene, 1) matrix}

\item{seeds}{character vector}
}
\description{
Returns a dataframe object with the stationary probability value and column indicating whether gene was a seed
}
