% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bin_presence}
\alias{bin_presence}
\title{Check Presence in Bins for Numbers 1-100}
\usage{
bin_presence(x)
}
\arguments{
\item{x}{Numeric vector. Values should be between 1 and 100.}
}
\value{
A named logical vector of length 10. Each element is \code{TRUE} if at least one
value in \code{x} falls into the corresponding bin, otherwise \code{FALSE}.
}
\description{
Given a numeric vector with values between 1 and 100, returns a logical vector
indicating whether at least one value falls into each bin of size 10
(i.e., 1-10, 11-20, ..., 91-100).
}
\examples{
vec <- c(3, 15, 27, 45, 58, 99)
bin_presence(vec)
}
