% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{binarize_matrices}
\alias{binarize_matrices}
\title{Binarize Multiple Matrices Based on Cutoffs}
\usage{
binarize_matrices(matrix_list, cutoff_list)
}
\arguments{
\item{matrix_list}{A list of matrices to be binarized}

\item{cutoff_list}{A list of cutoff values, one for each matrix}
}
\value{
A list of binarized matrices
}
\description{
Binarize Multiple Matrices Based on Cutoffs
}
\details{
This function takes a list of matrices and a corresponding list of cutoff values.
For each matrix, values less than or equal to the cutoff are set to 0, and values
greater than the cutoff are set to 1.
}
