% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_methods.R
\name{correlated_sigs}
\alias{correlated_sigs}
\title{Recontextualize seed signatures with correlation based neighbors}
\usage{
correlated_sigs(corr_mat, seeds, limit = 30)
}
\arguments{
\item{corr_mat}{Correlation Matrix}

\item{seeds}{Seed or List of seeds}

\item{limit}{Numeric indicating the number of genes to be returned}
}
\description{
Recontextualize seed signatures with correlation based neighbors
}
