% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{count_displaced_genes}
\alias{count_displaced_genes}
\title{Count number of displaced seeds
Note: This not commutative. Num_displaced is the number of genes that are from the original seed that are not in the recon.}
\usage{
count_displaced_genes(recon_sig, seed_sig)
}
\arguments{
\item{recon_sig}{Named list of genesets}

\item{seed_sig}{Named list of genesets}
}
\value{
Named list for which each entry is a list with two elements: displaced seed genes, and non-displaced seed genes
}
\description{
Count number of displaced seeds
Note: This not commutative. Num_displaced is the number of genes that are from the original seed that are not in the recon.
}
