% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{diag_zero}
\alias{diag_zero}
\title{Set Diagonal of a Matrix to Zero}
\usage{
diag_zero(matrix)
}
\arguments{
\item{matrix}{A square matrix.}
}
\value{
A matrix of the same dimensions as the input, with diagonal elements set to zero.
}
\description{
This function sets the diagonal elements of a matrix to zero.
}
