% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{jaccard_matrix}
\alias{jaccard_matrix}
\title{Create a Jaccard Similarity Matrix for Multiple Sets}
\usage{
jaccard_matrix(sets)
}
\arguments{
\item{sets}{A list of sets (vectors)}
}
\value{
A matrix of Jaccard similarity scores between all pairs of sets
}
\description{
Create a Jaccard Similarity Matrix for Multiple Sets
}
