% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{largest_connected_subgraph}
\alias{largest_connected_subgraph}
\title{Extract Largest Connected Subgraph}
\usage{
largest_connected_subgraph(igraph)
}
\arguments{
\item{igraph}{An igraph object}
}
\value{
An igraph object representing the largest connected component of the input graph
}
\description{
Extract Largest Connected Subgraph
}
