% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_methods.R
\name{network_sig_path}
\alias{network_sig_path}
\title{Generate Network Signatures from a Path}
\usage{
network_sig_path(
  path,
  seeds,
  sig = c("corr", "rwr"),
  avg_p = FALSE,
  avg_p_vals = c(1e-04, 0.1),
  avg_p_length = 5,
  p = 0.1,
  bootstrap = FALSE,
  limit = 30
)
}
\arguments{
\item{path}{A character string specifying the file path to the network object (RDS file).}

\item{seeds}{Either a single unnamed gene "TP53", a named list of genes, or a list of named lists of genes.}

\item{sig}{The signature generation method. Either "corr" for correlation-based or "rwr" for random walk with restart. Default is c("corr", "rwr").}

\item{p}{A numeric value specifying the restart probability for random walk. Default is 0.1.}

\item{bootstrap}{A boolean specifying whether to use empirical distributions of stationary values to find significant genes.}

\item{limit}{Number of genes to keep in the output, or a vector of lengths. Default is 30.}

\item{n_bootstraps}{A numeric specifying the number of bootstraps to perform.}

\item{n_cores}{A numeric indicating the number of cores for multi-core processing.}
}
\value{
A list of network signatures. If the input is a single network, returns a single signature list. If the input is a list of networks, returns a list of signature lists.
}
\description{
This function reads a network object from a file path and generates network signatures based on provided seeds.
}
