% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_methods.R
\name{projectCor}
\alias{projectCor}
\title{Reconstruct Gene Signatures Using Projection Scores}
\usage{
projectCor(se, sigs, score = c("gsva", "eigen"))
}
\arguments{
\item{se}{A SummarizedExperiment object containing gene expression data.}

\item{sigs}{A list of gene signatures, where each element is a character vector of gene names.}

\item{score}{Scoring method used to score samples against input signatures.
Either \code{"gsva"} or \code{"eigen"}. Default is \code{"gsva"}.}
}
\value{
A list of reconstructed gene signatures, with the same structure as the input \code{sigs}.
}
\description{
This function reconstructs gene signatures based on their correlation with
per-sample projection scores computed from the input signatures.
}
\details{
The function performs the following steps:
\enumerate{
\item Calculates projection scores for the input signatures with GSVA or eigengenes.
\item Computes the correlation between gene expression and projection scores.
\item Ranks genes based on their correlation with each signature's projection scores.
\item Selects the top-ranking genes to form new signatures of the same length as the original ones.
}
}
