% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_evaluation.R
\name{recon_eval_df}
\alias{recon_eval_df}
\title{Evaluate recontextualized signatures}
\usage{
recon_eval_df(
  ig,
  seed_name,
  source_sigs,
  dest_sigs,
  split = 0,
  experiment = c("CCLE", "perturb-seq", "drugmatrix", "sciplex"),
  restart = 0.75,
  recon = TRUE,
  use_weights = TRUE,
  weights.pwr = 1,
  normalize = c("row", "column", "laplacian"),
  limit = 30
)
}
\arguments{
\item{ig}{igraph}

\item{seed_name}{name of list of genesets to be used to label dataframe}

\item{source_sigs}{list of genesets (to be recontextualized)}

\item{dest_sigs}{list of genesets (the ground truth)}

\item{split}{split number for CCLE TCGA experiment}

\item{experiment}{benchmarking task type}

\item{restart}{restart value}

\item{recon}{Boolean indicating whether recontextualization occurs}

\item{use_weights}{Boolean indicating whether to use weights in the KS Test}

\item{weights.pwr}{Power to raise weights to}

\item{normalize}{Normalization strategy to employ}

\item{limit}{number of genes to be included in the network signature, default=30}
}
\description{
Evaluate recontextualized signatures
}
