% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_methods.R
\name{rwr_mat}
\alias{rwr_mat}
\title{Perform a random walk on an igraph given seeds, return stationary probabilities}
\usage{
rwr_mat(
  ig,
  seeds,
  restart = 0.75,
  avg_p = FALSE,
  avg_p_vals = c(1e-04, 0.1),
  avg_p_length = 5,
  epsilon = NULL,
  normalize = c("row", "column", "laplacian", "none")
)
}
\arguments{
\item{ig}{igraph object}

\item{seeds}{Either a single unnamed gene "TP53", a named list of genes, or a list of named lists of genes}

\item{restart}{A numeric specifying the probability of restarting at the seed nodes}

\item{avg_p}{A boolean specifying whether to ensemble random walk results over a range of restart values}

\item{avg_p_vals}{A numeric vector specifying the start and end of a arithmetic sequence to explore restart values}

\item{avg_p_length}{A numeric specifying how many values within \code{avg_p_vals} to include in the ensemble}

\item{epsilon}{Exploration factor}

\item{normalize}{Normalization strategy}
}
\value{
(n_gene, n_seeds) matrix of stationary probability values
}
\description{
Perform a random walk on an igraph given seeds, return stationary probabilities
}
