% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{seurat_common_var_genes}
\alias{seurat_common_var_genes}
\title{Find Common Variable Genes Across Seurat Objects}
\usage{
seurat_common_var_genes(seurat_objs, limit)
}
\arguments{
\item{seurat_objs}{A list of Seurat objects to compare.}

\item{limit}{An integer specifying the maximum number of common variable genes to return.}
}
\value{
A character vector of common variable gene names.
}
\description{
This function identifies common variable genes across multiple Seurat objects,
up to a specified limit.
}
\details{
The function iterates through the variable features of each Seurat object,
selecting genes that are present in all objects. It continues until it reaches
the specified limit or exhausts all common variable genes.
}
