% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{sft.check}
\alias{sft.check}
\title{Check and Return Soft-Thresholding Power from WGCNA}
\usage{
sft.check(sft)
}
\arguments{
\item{sft}{A list object returned by WGCNA's pickSoftThreshold or pickSoftThreshold.fromSimilarity function.}
}
\value{
An integer representing the soft-thresholding power to be used.
}
\description{
This function checks the soft-thresholding power estimate from WGCNA's scale-free topology analysis
and returns an appropriate power value.
}
\details{
If a valid power estimate is found in the input, it is returned. If the power estimate is NA,
a default value of 6 is returned. The function prints a message indicating which power is being used.
}
