% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_evaluation.R
\name{sig_eval_table}
\alias{sig_eval_table}
\title{Evaluate Signature Prediction}
\usage{
sig_eval_table(source_sigs, pred_sigs, true_sigs, source = "source_context")
}
\arguments{
\item{source_sigs}{A named list of genesets.}

\item{pred_sigs}{A named list of genesets.}

\item{true_sigs}{A named list of genesets containing both the cutoff signature \code{up} and the full signature \code{up_full}.}

\item{source}{A character string describing the starting biological context.}
}
\value{
A data frame with evaluation results. Each row corresponds to a perturbation and context.
}
\description{
This function calculates various evaluation metrics for predicted gene signatures
against true (ground truth) gene signatures, given a source signature as input.
Metrics include Jaccard index, gene displacement, and Kolmogorov-Smirnov (KS) test statistics
for rank agreement.
}
