% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_methods.R
\name{v.correlated_sigs}
\alias{v.correlated_sigs}
\title{Recontextualize seed signatures with correlation based neighbors}
\usage{
v.correlated_sigs(corr_mats, seeds, limit = 30)
}
\arguments{
\item{corr_mats}{List of correlation matrices}

\item{seeds}{List of seeds, length has to match corr_mats}

\item{limit}{Number of genes to keep in the output, or a vector of lengths. Default is 30.}
}
\description{
Recontextualize seed signatures with correlation based neighbors
}
