% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{v.jaccard}
\alias{v.jaccard}
\title{Calculate Jaccard Similarity for Multiple Pairs of Sets}
\usage{
v.jaccard(sigs1, sigs2)
}
\arguments{
\item{sigs1}{A named list of genesets (vectors)}

\item{sigs2}{A named list of genesets (vectors)}
}
\value{
A vector of Jaccard similarity scores
}
\description{
Calculate Jaccard Similarity for Multiple Pairs of Sets
}
